Footer and Header - SASS updates to Media Queries
You will notice that the header and footer are now working with new
design. Still need to update the mobile designs as I cannot locate old
CSS files from App to do this. HTML structure is set for further
updates to the SASS. Still need to complete Jobs Page and Contact us
Page.